pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsapp-zip-to-docx"
version = "0.1.0"
description = "Convert WhatsApp chat exports to Word documents."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-docx>=1.1.2",
  "google-api-python-client>=2.164.0",
  "google-auth-oauthlib>=1.2.1",
]

[project.scripts]
whatsapp-zip-to-docx = "whatsapp_zip_to_docx.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
